Change Tile-IDs to Integer-Values. This is just a tmp. fix, the lower- and upper layer need to be refactored for easier use.

+++ src/de/grufty/rpgreader/engine/LuciferMapUnit.java
@@ -18,8 +18,8 @@ public class LuciferMapUnit {
 	private long horizontalScrollSpeed		= 0;
 	private long verticalScrollSpeed		= 0;
 	private long timesSaved					= 0;
-	private ArrayList<Long> lowerLayer;
-	private ArrayList<Long> upperLayer;
+	private ArrayList<Integer> lowerLayer;
+	private ArrayList<Integer> upperLayer;
 	private boolean usePanorama				= false;
 	private boolean horizontalPan			= false;
 	private boolean verticalPan				= false;
@@ -96,16 +96,16 @@ public class LuciferMapUnit {
 				break;
 			case 0x47:
 				tmp = new DataReader(unit.content);
-				lowerLayer = new ArrayList<Long>(unit.content.length / 2);
+				lowerLayer = new ArrayList<Integer>(unit.content.length / 2);
 				for (int i = 0; i < unit.content.length / 2; i++) {
-					lowerLayer.add(tmp.next16bitle()); //TODO check if correct
+					lowerLayer.add((int) tmp.next16bitle()); //TODO check if correct
 				}
 				break;
 			case 0x48:
 				tmp = new DataReader(unit.content);
-				upperLayer = new ArrayList<Long>(unit.content.length / 2);
+				upperLayer = new ArrayList<Integer>(unit.content.length / 2);
 				for (int i = 0; i < unit.content.length / 2; i++) {
-					upperLayer.add(tmp.next16bitle()); //TODO check if correct
+					upperLayer.add((int) tmp.next16bitle()); //TODO check if correct
 				}
 				break;
 			case 0x51:
@@ -266,7 +266,7 @@ public class LuciferMapUnit {
 	 * 
 	 * @return the lowerLayer
 	 */
-	public ArrayList<Long> getLowerLayer() {
+	public ArrayList<Integer> getLowerLayer() {
 		return lowerLayer;
 	}
 
@@ -276,7 +276,7 @@ public class LuciferMapUnit {
 	 * @param lowerLayer the new lowerLayer
 	 */
 	public void setLowerLayer(
-			ArrayList<Long> lowerLayer) {
+			ArrayList<Integer> lowerLayer) {
 		this.lowerLayer = lowerLayer;
 	}
 
@@ -285,7 +285,7 @@ public class LuciferMapUnit {
 	 * 
 	 * @return the upperLayer
 	 */
-	public ArrayList<Long> getUpperLayer() {
+	public ArrayList<Integer> getUpperLayer() {
 		return upperLayer;
 	}
 
@@ -295,7 +295,7 @@ public class LuciferMapUnit {
 	 * @param upperLayer the new upperLayer
 	 */
 	public void setUpperLayer(
-			ArrayList<Long> upperLayer) {
+			ArrayList<Integer> upperLayer) {
 		this.upperLayer = upperLayer;
 	}
